requirements.txt
Django==2.1.4
django-crispy-forms==1.7.2
docutils==0.14
lockfile==0.12.2
psycopg2==2.7.6.1
pytz==2018.7
RPi.GPIO==0.6.5
serpent==1.27
smbus2==0.2.1
